requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
ecdsa==0.18.0
idna==3.4
Markdown==3.5.1
mysql-connector-python==8.2.0
mysqlclient==2.2.0
Pillow==10.1.0
protobuf==4.21.12
pyasn1==0.5.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
pythonping==1.1.4
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
